+++ empire/server/modules/powershell/credentials/credential_injection.py
@@ -1,33 +1,27 @@
 from empire.server.common.empire import MainMenu
 from empire.server.core.db.base import SessionLocal
+from empire.server.core.exceptions import ModuleValidationException
 from empire.server.core.module_models import EmpireModule
-from empire.server.utils.module_util import handle_error_message
+from empire.server.core.module_service import auto_finalize, auto_get_source
 
 
 class Module:
     @staticmethod
+    @auto_get_source
+    @auto_finalize
     def generate(
         main_menu: MainMenu,
         module: EmpireModule,
         params: dict,
         obfuscate: bool = False,
         obfuscation_command: str = "",
+        script: str = "",
     ):
-        # read in the common module source code
-        script, err = main_menu.modulesv2.get_module_source(
-            module_name=module.script_path,
-            obfuscate=obfuscate,
-            obfuscate_command=obfuscation_command,
-        )
-
-        if err:
-            return handle_error_message(err)
-
         script_end = "Invoke-CredentialInjection"
 
         if params["NewWinLogon"] == "" and params["ExistingWinLogon"] == "":
-            return handle_error_message(
-                "[!] Either NewWinLogon or ExistingWinLogon must be specified"
+            raise ModuleValidationException(
+                "Either NewWinLogon or ExistingWinLogon must be specified"
             )
 
         # if a credential ID is specified, try to parse
@@ -37,11 +31,11 @@ class Module:
                 cred = main_menu.credentialsv2.get_by_id(db, cred_id)
 
                 if not cred:
-                    return handle_error_message("[!] CredID is invalid!")
+                    raise ModuleValidationException("CredID is invalid")
 
                 if cred.credtype != "plaintext":
-                    return handle_error_message(
-                        "[!] A CredID with a plaintext password must be used!"
+                    raise ModuleValidationException(
+                        "CredID must be a plaintext credential"
                     )
 
                 if cred.domain != "":
@@ -56,8 +50,8 @@ class Module:
             or params["UserName"] == ""
             or params["Password"] == ""
         ):
-            return handle_error_message(
-                "[!] DomainName/UserName/Password or CredID required!"
+            raise ModuleValidationException(
+                "DomainName/UserName/Password or CredID required"
             )
 
         for option, values in params.items():
@@ -69,10 +63,4 @@ class Module:
                     else:
                         script_end += " -" + str(option) + " " + str(values)
 
-        script = main_menu.modulesv2.finalize_module(
-            script=script,
-            script_end=script_end,
-            obfuscate=obfuscate,
-            obfuscation_command=obfuscation_command,
-        )
-        return script
+        return script, script_end